backend/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
motor>=3.6,<4
pydantic>=2.8,<3
pydantic-settings>=2.4,<3
email-validator>=2.2,<3
python-dotenv>=1.0,<2
python-multipart>=0.0.9,<1
passlib[bcrypt]>=1.7,<2
python-jose[cryptography]>=3.3,<4